requirements.txt
fastapi==0.115.0
motor==3.6.0
pydantic==2.11.7
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pymongo[ocsp]==4.8.0  # Required for Motor, includes OCSP for SSL